+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add Okta provider
 * Add Cidaas provider
 * Add help link to `heptacom-admin-open-auth-client-edit-page`
+* Add support verify using OAuth when asked for a password to verify in `sw-verify-user-modal`
 
 ## Changed
 * Changed create provider page to get a better overview of the existing providers